post

+++ ToDolistBack/src/main/java/com/doido/todolistback/domain/task/servicies/impl/TaskServiceImple.java
@@ -1,18 +1,32 @@
 package com.doido.todolistback.domain.task.servicies.impl;
 
-import com.doido.todolistback.domain.task.dtos.post.PostTaskDto;
-import com.doido.todolistback.domain.task.dtos.request.RequestTaskDto;
+import com.doido.todolistback.domain.task.dtos.request.SubTaskRequest;
+import com.doido.todolistback.domain.task.dtos.response.SubTaskResponse;
+import com.doido.todolistback.domain.task.dtos.response.TaskResponse;
+import com.doido.todolistback.domain.task.entity.SubTask;
 import com.doido.todolistback.domain.task.entity.Task;
 import com.doido.todolistback.domain.task.servicies.TaskService;
 import com.doido.todolistback.domain.task.shared.mapper.TaskMapper;
-import com.doido.todolistback.infra.repositories.TaskRepository;
+import com.doido.todolistback.domain.task.specification.TaskSpecification;
+import com.doido.todolistback.domain.user.entity.User;
+import com.doido.todolistback.domain.task.repositories.TaskRepository;
 import com.doido.todolistback.shared.exception.CustomsExceptions.TaskNotFound;
+import com.doido.todolistback.shared.exception.CustomsExceptions.UserNotFoundException;
+import com.doido.todolistback.shared.utils.SecurityUtils;
 import jakarta.transaction.Transactional;
+import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.UUID;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 @Transactional
@@ -23,9 +37,10 @@ public class TaskServiceImple implements TaskService {
 
 
     @Override
-    public PostTaskDto addTask(PostTaskDto task){
+    public com.doido.todolistback.domain.task.dtos.request.TaskRequest addTask(com.doido.todolistback.domain.task.dtos.request.TaskRequest task){
+        Task taskEntity = taskMapper.postTaskDtoToTask(task);
 
-        Task taskEntity = taskMapper.toTaskPost(task);
+        taskEntity.setUser((User) SecurityUtils.getUser());
         taskEntity.setCompleted(false);
 
         taskRepository.save(taskEntity);
@@ -33,10 +48,19 @@ public class TaskServiceImple implements TaskService {
     }
 
     @Override
-    public PostTaskDto updateTask(Long id, PostTaskDto task){
+    public com.doido.todolistback.domain.task.dtos.request.TaskRequest updateTask(UUID id, com.doido.todolistback.domain.task.dtos.request.TaskRequest task){
 
-        Task taskEntity = taskRepository.findById(id)
-                            .orElseThrow(()->new TaskNotFound());
+        User user = (User) SecurityUtils.getUser();
+
+        Task taskEntity = taskRepository
+                .findById(id)
+                .orElseThrow(()->new TaskNotFound());
+
+
+        if (!taskEntity.getUser().equals(user)){
+            throw new UserNotFoundException("");
+        }
+
 
         if (task.getTitle() != null && !task.getTitle().equals("")){
             taskEntity.setTitle(task.getTitle());
@@ -47,23 +71,108 @@ public class TaskServiceImple implements TaskService {
 
         taskRepository.save(taskEntity);
 
-        return taskMapper.toPostTaskDto(taskEntity);
+        return taskMapper.taskToPostTaskDto(taskEntity);
+
+
+
+
     }
 
     @Override
-    public void deleteTask(Long id){
-        taskRepository.findById(id).orElseThrow(() -> new TaskNotFound());
-        taskRepository.deleteById(id);
+    public TaskResponse statusTask(boolean status, UUID TaskId){
+        Task task = isOwner(TaskId);
+        task.setCompleted(status);
+        taskRepository.save(task);
+
+        return taskMapper.taskResponse(task);
     }
 
     @Override
-    public List<RequestTaskDto> findAll(){
-        List<Task> task = taskRepository.findAll();
+    public void deleteTask(UUID id){
 
-        List<RequestTaskDto> requestTaskDto = task.stream()
-            .map(taskMapper::toRequestTaskDto)
+        User user = (User) SecurityUtils.getUser();
+
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new TaskNotFound());
+
+        if (task.getUser().getId().equals(user.getId())){
+            taskRepository.deleteById(id);
+        } else {throw new UserNotFoundException("The user is not the owner of the task.");}
+
+
+
+    }
+
+    @Override
+    public List<TaskResponse> findUserTask(){
+
+        User user = (User) SecurityUtils.getUser();
+
+        List<Task> task = taskRepository.findByUserId(user.getId());
+
+        List<TaskResponse> taskResponse = task.stream()
+            .map(taskMapper::taskResponse)
             .toList();
 
-        return requestTaskDto;
+        return taskResponse;
+    }
+
+    @Override
+    public Page<TaskResponse> filter(Boolean completed, Pageable pageable) {
+
+        User user = (User) SecurityUtils.getUser();
+
+        Specification spec = Specification
+                .where(TaskSpecification.completed(completed));
+
+
+        return taskRepository.findAllByUserId(user.getId(), pageable, spec)
+                .map(taskMapper::taskResponse);
+    }
+
+    //SubTask
+
+    @Override
+    public SubTaskResponse addSubTask(SubTaskRequest subTask, UUID idTarefa) {
+
+            Task task = isOwner(idTarefa);
+
+            SubTask subTaskEntity = taskMapper.postDtotoSubTask(subTask);
+
+            task.addSubTask(subTaskEntity);
+
+            taskRepository.save(task);
+
+            return  taskMapper.subTaskToRequestDto(subTaskEntity);
+    }
+
+
+
+    @Override
+    @Transactional
+    public List<SubTaskResponse> findSubTarefa(UUID id) {
+
+        Task task = isOwner(id);
+
+        return task.getSubTasks()
+                .stream()
+                .map(taskMapper::subTaskToRequestDto)
+                .toList();
+    }
+
+
+
+    //Private Method
+
+    private Task isOwner(UUID id){
+        User user = (User) SecurityUtils.getUser();
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new TaskNotFound());
+
+        if (!user.getId().equals(task.getUser().getId())){
+            throw new RuntimeException("User is not the owner of the task.");
+        }
+
+        return task;
     }
 }
